Moved the surface formatter for general use - maybe should even be renamed
